Added new brushes and reset the default brush
Many brushes are just scaled versions of a different brush.  I think that more brush variety should be had, and if a brush needs to be scaled, use the Tool Size slider/property

The code for creating a circle is a little off, because the circle top and bottom is cut.  Since this is a brush, and should show variety, I haven't "fixed" this code yet.

Note that
the fuzzy circle brush act like an airspray if Continuous Mode is disabled.
the circle outline brush has a better effect also if Continous Mode is disabled